clean up deepspeech refactoring
